+++ next/src/utils/loginLoader.ts
@@ -2,10 +2,15 @@ import { redirect } from '@tanstack/react-router';
 
 import { getOidc } from '@/lib/auth/oidc';
 
+import { getCurrentUri } from './utils';
+
 export async function loginLoader() {
-  const oidc = await getOidc();
-  if (!oidc.isUserLoggedIn) {
-    throw redirect({ to: '/login' });
+  const { isUserLoggedIn } = await getOidc();
+  if (!isUserLoggedIn) {
+    throw redirect({
+      to: '/login',
+      search: { redirectUri: getCurrentUri() },
+    });
   }
   return null;
 }