+++ next/src/utils/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
 
-import { getUrlFromCriteria, nameFromLabel } from './utils';
+import { getCurrentUri, getUrlFromCriteria, nameFromLabel } from './utils';
 
 describe('Utils', () => {
   test.each([
@@ -23,5 +23,51 @@ describe('Utils', () => {
     ['mon 1er läbél', 'MON1ERLBL'],
   ])('nameFromLabel (%s) -> %s', (label, expected) => {
     expect(nameFromLabel(label)).toEqual(expected);
+  });
+});
+
+describe('getCurrentUri', () => {
+  const originalLocation = globalThis.location;
+
+  const mockLocation = (href: string, origin: string) => {
+    Object.defineProperty(globalThis, 'location', {
+      value: { href, origin },
+    });
+  };
+
+  afterEach(() => {
+    Object.defineProperty(globalThis, 'location', {
+      value: originalLocation,
+    });
+  });
+
+  it('should return pathname', () => {
+    mockLocation('https://example.com/questionnaire', 'https://example.com');
+
+    expect(getCurrentUri()).toBe('/questionnaire');
+  });
+
+  it('should return pathname with query params', () => {
+    mockLocation(
+      'https://example.com/questionnaire?param=2',
+      'https://example.com',
+    );
+
+    expect(getCurrentUri()).toBe('/questionnaire?param=2');
+  });
+
+  it('should return pathname with query params and hash', () => {
+    mockLocation(
+      'https://example.com/questionnaire?param=2#hash',
+      'https://example.com',
+    );
+
+    expect(getCurrentUri()).toBe('/questionnaire?param=2#hash');
+  });
+
+  it('should return "/" for root', () => {
+    mockLocation('https://example.com/', 'https://example.com');
+
+    expect(getCurrentUri()).toBe(undefined);
   });
 });