+++ src/auth/__mocks__/context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+export const AuthContext = createContext({
+  getAccessToken: () => Promise.resolve(undefined),
+  decodedIdToken: {
+    timbre: 'my-stamp',
+  },
+});