+++ src/layout/page-search-questionnaire/components/search-results/index.jsx
@@ -1,2 +1,0 @@
-export { default as SearchResults } from './search-results-container';
-export { defaultProps } from './search-results-container';