+++ src/layout/page-search-questionnaire/index.jsx
@@ -1,1 +1,0 @@
-export { default as PageSearchQuestionnaire } from './page-search-questionnaire-container';