+++ next/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- restore desired uri after login
+
+### Changed
+
+- migrate to latest oidc-spa (v10)
+- add `getAccessToken` & `decodedIdToken` to params of legacy App
+
 ## [2.5.0](https://github.com/InseeFr/Pogues/releases/tag/2.5.0) - 2026-03-06
 
 ### Changed